# swagger-api/openapi.yaml
openapi: 3.0.3
info:
  title: Cherry Deploy Backend API
  version: 0.1.0
  description: |
    Unified specification combining chatbot and DevOps endpoints. Requires the live FastAPI backend for
    actual execution because several endpoints depend on background tasks, local repositories, and MongoDB.
servers:
  - url: https://delight.13-125-116-92.nip.io
    description: Production-like endpoint (HTTPS via Let's Encrypt)
  - url: http://127.0.0.1:9001
    description: Local FastAPI server (bypasses Nginx)
paths:
  /api/v1/chat:
    post:
      summary: Send a user message and receive a Gemini-generated reply.
      tags: [chat]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
            examples:
              basic:
                summary: Basic greeting
                value:
                  message: "안녕하세요! 배포 상황이 어떤지 알려주세요."
      responses:
        '200':
          description: Gemini replied successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Invalid prompt payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/deploy:
    post:
      summary: Trigger a chatbot frontend deployment
      tags: [deploy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeployRequest'
            example:
              branch: deploy
      responses:
        '202':
          description: Deployment accepted; background task scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployResponse'
        '500':
          description: Server error prevented task creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/status/{task_id}:
    get:
      summary: Fetch current deployment status
      tags: [deploy]
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task found; returns status and metadata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployStatusResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/preview:
    get:
      summary: Describe upcoming deploy steps, risks, and costs
      tags: [deploy]
      parameters:
        - in: query
          name: task_id
          schema:
            type: string
          description: Optional task identifier to align the preview timeline with a specific run.
      responses:
        '200':
          description: Static preview payload showing operational context.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployPreviewResponse'
  /api/v1/rollback:
    post:
      summary: Roll back to the previous successful deployment commit
      tags: [deploy]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackRequest'
      responses:
        '202':
          description: Rollback accepted; background task scheduled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployResponse'
        '400':
          description: Invalid rollback request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Not enough history to perform rollback.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/tasks/recent:
    get:
      summary: List recent deploy/rollback tasks
      tags: [deploy]
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 5
          description: Maximum number of recent tasks to return (clamped to 20).
      responses:
        '200':
          description: Recent tasks ordered by start time (newest first).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeployTaskSummary'
  /api/v1/tasks/{task_id}/logs:
    get:
      summary: Detailed stdout/stderr metadata for a specific task
      tags: [deploy]
      parameters:
        - in: path
          name: task_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task metadata with captured stdout/stderr snippets.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeployTaskLogResponse'
        '404':
          description: Task not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /healthz:
    get:
      summary: Liveness probe
      tags: [system]
      responses:
        '200':
          description: Service alive indicator including PM2 and Mongo checks.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatusResponse'
components:
  schemas:
    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          minLength: 1
          description: User message forwarded to Gemini 2.5 Flash.
    ChatResponse:
      type: object
      required: [reply, model]
      properties:
        reply:
          type: string
        model:
          type: string
    DeployRequest:
      type: object
      properties:
        branch:
          type: string
          description: "Optional override for the deployment branch (default: deploy)."
          example: deploy
    DeployResponse:
      type: object
      required:
        - task_id
        - status
        - branch
        - action
        - queued_at
        - estimated_duration_minutes
        - context
        - dev_server_restart_planned
      properties:
        task_id:
          type: string
        status:
          $ref: '#/components/schemas/DeployStatusEnum'
        branch:
          type: string
          description: Branch resolved for this deployment or rollback.
        action:
          type: string
          enum: [deploy, rollback]
        queued_at:
          type: string
          format: date-time
        estimated_duration_minutes:
          type: integer
          minimum: 1
          description: Rough ETA derived from diff statistics.
        context:
          type: object
          additionalProperties: true
        dev_server_restart_planned:
          type: boolean
          description: True when the pipeline restarts the long-running dev server instead of copying static assets.
    DeployStatusResponse:
      type: object
      required: [task_id, status, metadata, stages, started_at]
      properties:
        task_id:
          type: string
        status:
          $ref: '#/components/schemas/DeployStatusEnum'
        metadata:
          type: object
          additionalProperties: true
        stages:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error_log:
          type: string
          nullable: true
        failure_context:
          type: object
          nullable: true
          additionalProperties: true
        cost_estimate:
          type: object
          additionalProperties: true
          nullable: true
        risk_assessment:
          type: object
          additionalProperties: true
          nullable: true
        llm_preview:
          $ref: '#/components/schemas/LLMPreview'
          nullable: true
        blue_green_plan:
          $ref: '#/components/schemas/BlueGreenPlan'
          nullable: true
    DeployPreviewResponse:
      type: object
      required:
        - current_branch
        - target_repo
        - commands
        - risk_assessment
        - cost_estimate
        - timeline_preview
        - warnings
        - blue_green_plan
      properties:
        current_branch:
          type: string
        target_repo:
          type: string
        frontend_project_path:
          type: string
          nullable: true
        frontend_output_path:
          type: string
          nullable: true
        commands:
          type: array
          items:
            type: string
          example:
            - "git fetch origin"
            - "git checkout -B deploy origin/deploy"
            - "git reset --hard origin/deploy"
            - "git clean -fdx"
            - "npm install"
            - "bash -lc pm2 delete frontend-dev 2>/dev/null || true; pm2 start npm --name frontend-dev -- run dev -- --hostname 0.0.0.0 --port 3000"
            - "no static cutover (dev-server mode)"
        risk_assessment:
          type: object
          additionalProperties: true
        cost_estimate:
          type: object
          additionalProperties: true
        llm_preview:
          $ref: '#/components/schemas/LLMPreview'
          nullable: true
        timeline_preview:
          type: array
          items:
            $ref: '#/components/schemas/DeployTimelineEntry'
        warnings:
          type: array
          items:
            type: string
        task_context:
          $ref: '#/components/schemas/DeployTaskSummary'
          nullable: true
        blue_green_plan:
          $ref: '#/components/schemas/BlueGreenPlan'
    RollbackRequest:
      type: object
      properties:
        branch:
          type: string
          description: Optional override for repo1 branch (defaults to deploy).
          example: deploy
    DeployStatusEnum:
      type: string
      enum:
        - pending
        - running_clone
        - running_build
        - running_cutover
        - running_observability
        - completed
        - failed
    ErrorResponse:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    DeployTimelineEntry:
      type: object
      required: [stage, label, completed]
      properties:
        stage:
          $ref: '#/components/schemas/DeployStatusEnum'
        label:
          type: string
        expected_seconds:
          type: integer
          minimum: 0
          nullable: true
        completed:
          type: boolean
        metadata:
          type: object
          nullable: true
          additionalProperties: true
    DeployTaskSummary:
      type: object
      required:
        - task_id
        - status
        - branch
        - action
        - started_at
      properties:
        task_id:
          type: string
        status:
          $ref: '#/components/schemas/DeployStatusEnum'
        branch:
          type: string
        action:
          type: string
          description: Operation type (deploy or rollback).
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        summary:
          type: object
          nullable: true
          additionalProperties: true
        failure_context:
          type: object
          nullable: true
          additionalProperties: true
    DeployTaskLogResponse:
      type: object
      required:
        - task_id
        - status
        - stages
        - metadata
      properties:
        task_id:
          type: string
        status:
          $ref: '#/components/schemas/DeployStatusEnum'
        stages:
          type: object
          additionalProperties:
            type: object
            additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
        error_log:
          type: string
          nullable: true
        failure_context:
          type: object
          nullable: true
          additionalProperties: true
    HealthStatusResponse:
      type: object
      required: [status, pm2_processes, mongo, issues]
      properties:
        status:
          type: string
          description: "healthy when all checks pass, otherwise degraded."
        pm2_processes:
          type: object
          additionalProperties:
            type: string
          description: Map of PM2 process names to their states.
        mongo:
          type: string
          description: Result of the MongoDB ping command.
        last_task_id:
          type: string
          nullable: true
        last_task_status:
          $ref: '#/components/schemas/DeployStatusEnum'
          nullable: true
        issues:
          type: array
          items:
            type: string
        blue_green:
          $ref: '#/components/schemas/BlueGreenPlan'
    LLMPreview:
      type: object
      required: [summary, highlights, risks]
      properties:
        summary:
          type: string
        highlights:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
    BlueGreenPlan:
      type: object
      required: [active_slot, next_cutover_target]
      properties:
        active_slot:
          type: string
          description: Currently serving color (green/blue/dev-server/unknown).
        standby_slot:
          type: string
          nullable: true
          description: Idle slot that will be served next if a cutover occurs.
        last_cutover_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent completed cutover, if known.
        next_cutover_target:
          type: string
          nullable: true
          description: Slot targeted for the next cutover step.